Working ShortCut GistImporter
Created a static class to allow the use of  the Gist Importer with the use of systemcopybuffer (i.e the stuff in the copy and past clipboard). Creating an easier way for users to implement Gists into the project with less hassle. The reason a static class was chosen instead of a menu item was to ensure no menu item was created and in turn, confuse users with more dropdown options